move CidVariant to utils::cid module
