+++ src/utils/cid/mod.rs
@@ -3,8 +3,9 @@
 
 use cid::{
     multihash::{Code::Blake2b256, MultihashDigest},
-    Cid,
+    Cid, Version,
 };
+use fvm_ipld_encoding::DAG_CBOR;
 
 /// Extension methods for constructing `dag-cbor` [Cid]
 pub trait CidCborExt {
@@ -25,6 +26,30 @@ pub trait CidCborExt {
 
 impl CidCborExt for Cid {}
 
+const BLAKE2B256: u64 = 0xb220;
+pub const BLAKE2B256_SIZE: usize = 32;
+
+// Nearly all Filecoin CIDs are V1, DagCbor encoded, and hashed with Blake2b265. If other types of
+// CID become popular, they can be added to the CidVariant structure.
+pub enum CidVariant {
+    V1DagCborBlake2b([u8; BLAKE2B256_SIZE]),
+}
+
+impl TryFrom<Cid> for CidVariant {
+    type Error = ();
+    fn try_from(cid: Cid) -> Result<Self, Self::Error> {
+        if cid.version() == Version::V1 && cid.codec() == DAG_CBOR {
+            if let Ok(small_hash) = cid.hash().resize() {
+                let (code, bytes, size) = small_hash.into_inner();
+                if code == BLAKE2B256 && size as usize == BLAKE2B256_SIZE {
+                    return Ok(CidVariant::V1DagCborBlake2b(bytes));
+                }
+            }
+        }
+        Err(())
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -43,5 +68,58 @@ mod tests {
         ensure!(cid1 == cid2);
 
         Ok(())
+    }
+
+    use super::CidVariant;
+    use cid::{
+        multihash::{Code, MultihashDigest},
+        Cid,
+    };
+    use fvm_ipld_encoding::DAG_CBOR;
+    use std::mem::size_of;
+
+    // If this stops being true, please update the documentation above.
+    #[test]
+    fn cid_size_assumption() {
+        assert_eq!(size_of::<Cid>(), 96);
+    }
+
+    // If this stops being true, please update the BLAKE2B256 constant.
+    #[test]
+    fn blake_code_assumption() {
+        assert_eq!(Code::Blake2b256.digest(&[]).code(), super::BLAKE2B256);
+    }
+
+    // If this stops being true, please update the BLAKE2B256_SIZE constant.
+    #[test]
+    fn blake_size_assumption() {
+        assert_eq!(
+            Code::Blake2b256.digest(&[]).size() as usize,
+            super::BLAKE2B256_SIZE
+        );
+    }
+
+    #[test]
+    fn known_v1_blake2b() {
+        let cid = Cid::new(
+            cid::Version::V1,
+            DAG_CBOR,
+            Code::Blake2b256.digest("blake".as_bytes()),
+        )
+        .unwrap();
+        assert!(matches!(
+            cid.try_into().unwrap(),
+            CidVariant::V1DagCborBlake2b(_)
+        ));
+    }
+
+    // If this test fails, the default encoding is no longer v1+dagcbor.
+    #[test]
+    fn default_is_v1_dagcbor() {
+        let cid = Cid::from_cbor_blake2b256(&()).unwrap();
+        assert!(matches!(
+            cid.try_into().unwrap(),
+            CidVariant::V1DagCborBlake2b(_)
+        ));
     }
 }
